split out logic from websocket_server into other modules for connection lifecycle and cache keys, update tests
